Create README.md
Tests the conjecture that all lexicographically sorted matrices for any set x = 0 to (base b ^ n) – 1 where x, b and n are integers and b and n > 1 and <= 10 are comprised entirely of sets of rotated sequences that are in themselves or in pairs, palindromically isomorphic about the matrix median. Also tests that the same holds true for for all the permutations of any set base b where b  > 1 and <= 10.